Run zandronum-server as non-root user
Inside the container, a user and group is created which is used to run the
zandronum-server process. Environment variables have been introduced to allow
specifying the corresponding UID and GID (from the host machine) which will
be assigned to this local user and group. See the README for more detail.