Move CI to Github Actions & Support ARM
Also upgraded TSPG to v21 (from v20).

A patch file is provided for the lemon tool. According to the TSPG
developers, the version of lemon in 3.0.1 (or v21) does not function
properly on ARM (it segfaults). I did a diff between 3.0.1 in the
official Zandronum code base to the tip as of 09/17/2020 and saved that
as a patch. Note that the patch command failed to apply it until I
encoded the file as UTF-8, so that is essential going forward.